requirements.txt
tensorflow
keras
opencv-python
pandas
matplotlib
ultralytics
cvzone
gTTS
nltk
Pillow
tabulate
streamlit
